requirements.txt
config==0.4.2
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
django-bootstrap3==11.0.0
django-bootstrap4==0.0.8
django-contrib-comments==1.9.1
django-tinymce==2.8.0
djangorestframework==3.9.4
numpy==1.16.4
Pillow==6.0.0
psycopg2==2.8.2
python-decouple==3.1
pytz==2019.1
